refactor: extract FixCard as shared component and export from sharedComponents
FixCard was inline inside FixPanel; extracted to its own file with an
onSave callback prop instead of the hardcoded setPendingRescan dependency.
FixPanel updated to import and use it. FixCard exported from sharedComponents
so the pro plugin can render inline fix settings in the instance drawer.
