+++ src/sharedComponents/index.js
@@ -15,6 +15,7 @@ export {
 	getDismissReasonDescription,
 } from '../sidebar/utils/dismissHelpers';
 export { default as DismissPanel } from '../issueModal/components/DismissPanel';
+export { default as FixCard } from '../issueModal/components/FixCard';
 export { default as RichTextarea } from '../issueModal/components/RichTextarea';
 export { default as IssueImage, extractImageUrls } from '../issueModal/components/IssueImage';
 export { getSeverityLabel } from '../sidebar/utils/severityHelpers';
